refactor: remove dead code, consolidate CompanyLogo, and centralize interview constants
- Delete no-op `computeDateUpdates` from jobUtils.ts and its 8 tests
- Replace two local CompanyLogo implementations (RadarPage, InterviewsPage)
  with the canonical CompanyLogo component; update their specs to mock
  useCompanyLogo instead of logoCache
- Move INTERVIEW_STAGE_LABELS and INTERVIEW_TYPE_LABELS from InterviewsTab
  and InterviewsPage into constants.ts, eliminating the duplication
