+++ frontend/src/App.spec.tsx
@@ -2,7 +2,6 @@ import React from "react";
 import { act, render, screen, waitFor } from "@testing-library/react";
 import App from "./App";
 import { api, setUnauthorizedHandler } from "./api";
-import { computeDateUpdates } from "./jobUtils";
 import type { User } from "./types";
 
 let capturedUnauthorizedHandler: (() => void) | null = null;
@@ -119,62 +118,5 @@ describe(App, () => {
 				screen.getByRole("link", { name: "Continue with Google" }),
 			).toBeInTheDocument();
 		});
-	});
-});
-
-describe(computeDateUpdates, () => {
-	const NOW = "2026-03-25T14:30";
-	const jobWithDates = {
-		date_last_onsite: "2026-03-23T09:00",
-		date_phone_screen: "2026-03-20T10:00",
-	};
-	const jobNoDates = { date_last_onsite: null, date_phone_screen: null };
-
-	it("preserves both dates when moving to Phone screen", () => {
-		const result = computeDateUpdates(jobWithDates, "Phone screen", NOW);
-		expect(result.date_phone_screen).toBe(jobWithDates.date_phone_screen);
-		expect(result.date_last_onsite).toBe(jobWithDates.date_last_onsite);
-	});
-
-	it("preserves null dates when moving to Phone screen with no prior dates", () => {
-		const result = computeDateUpdates(jobNoDates, "Phone screen", NOW);
-		expect(result.date_phone_screen).toBeNull();
-		expect(result.date_last_onsite).toBeNull();
-	});
-
-	it("preserves both dates when moving to Interviewing", () => {
-		const result = computeDateUpdates(jobWithDates, "Interviewing", NOW);
-		expect(result.date_phone_screen).toBe(jobWithDates.date_phone_screen);
-		expect(result.date_last_onsite).toBe(jobWithDates.date_last_onsite);
-	});
-
-	it("preserves null dates when moving to Interviewing with no prior dates", () => {
-		const result = computeDateUpdates(jobNoDates, "Interviewing", NOW);
-		expect(result.date_phone_screen).toBeNull();
-		expect(result.date_last_onsite).toBeNull();
-	});
-
-	it("preserves both dates when moving back to Not started", () => {
-		const result = computeDateUpdates(jobWithDates, "Not started", NOW);
-		expect(result.date_phone_screen).toBe(jobWithDates.date_phone_screen);
-		expect(result.date_last_onsite).toBe(jobWithDates.date_last_onsite);
-	});
-
-	it("preserves both dates when moving back to Applied", () => {
-		const result = computeDateUpdates(jobWithDates, "Applied", NOW);
-		expect(result.date_phone_screen).toBe(jobWithDates.date_phone_screen);
-		expect(result.date_last_onsite).toBe(jobWithDates.date_last_onsite);
-	});
-
-	it("preserves both dates when moving to Offer!", () => {
-		const result = computeDateUpdates(jobWithDates, "Offer!", NOW);
-		expect(result.date_phone_screen).toBe(jobWithDates.date_phone_screen);
-		expect(result.date_last_onsite).toBe(jobWithDates.date_last_onsite);
-	});
-
-	it("preserves both dates when moving to Rejected/Withdrawn", () => {
-		const result = computeDateUpdates(jobWithDates, "Rejected/Withdrawn", NOW);
-		expect(result.date_phone_screen).toBe(jobWithDates.date_phone_screen);
-		expect(result.date_last_onsite).toBe(jobWithDates.date_last_onsite);
 	});
 });
